+++ src/components/index.ts
@@ -2,3 +2,4 @@ export { default as BButton } from './button/Button.vue';
 export { default as BContainer } from './grid/Container.vue';
 export { default as BGrid } from './grid/Grid.vue';
 export { default as BItem } from './grid/Item.vue';
+export { default as BDataTable } from './data-table/DataTable.vue';